Applied Comments. Example reworked.
