+++ examples/microprofile/bean-validation/src/main/resources/logging.properties
@@ -28,13 +28,3 @@ java.util.logging.SimpleFormatter.format=%1$tY.%1$tm.%1$td %1$tH:%1$tM:%1$tS %4$
 
 # Quiet Weld
 org.jboss.level=WARNING
-
-# Component specific log levels
-#io.helidon.webserver.level=INFO
-#io.helidon.config.level=INFO
-#io.helidon.security.level=INFO
-#io.helidon.microprofile.level=INFO
-#io.helidon.common.level=INFO
-#io.netty.level=INFO
-#org.glassfish.jersey.level=INFO
-#org.jboss.weld=INFO
