--- tests/integration/mp-bean-validation/src/main/java/io/helidon/tests/integration/bean/validation/GreetingHolder.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2022 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.tests.integration.bean.validation;
-
-
-import jakarta.validation.constraints.NotNull;
-
-/**
- * Holder for a greeting, ensuring it is not null.
- */
-public class GreetingHolder {
-
-    /**
-     * Explicitly declare as not eligible for null.
-     */
-    @NotNull
-    private String greeting;
-
-    GreetingHolder(String greeting) {
-        this.greeting = greeting;
-    }
-
-    String getGreeting() {
-        return greeting;
-    }
-
-}
-
