--- tests/integration/mp-bean-validation/src/main/java/io/helidon/tests/integration/bean/validation/GreetingProvider.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2022 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.helidon.tests.integration.bean.validation;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-import jakarta.validation.Valid;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-/**
- * Provider for greeting message.
- */
-@ApplicationScoped
-public class GreetingProvider {
-
-    /**
-     * Explicitly declare that greeting should not be null.
-     */
-    private GreetingHolder greetingHolder;
-
-    /**
-     * Create a new greeting provider, reading the message from configuration.
-     *
-     * @param message greeting to use
-     */
-    @Inject
-    public GreetingProvider(@ConfigProperty(name = "app.greeting") String message) {
-        this.greetingHolder = new GreetingHolder(message);
-    }
-
-    String getGreeting() {
-        return greetingHolder.getGreeting();
-    }
-
-    void setGreeting(@Valid GreetingHolder greetingHolder) {
-        this.greetingHolder = greetingHolder;
-    }
-}
