+++ tests/integration/mp-bean-validation/src/main/resources/META-INF/microprofile-config.properties
@@ -14,9 +14,6 @@
 # limitations under the License.
 #
 
-# Application properties. This is the default greeting
-app.greeting=Hello
-
 # Microprofile server properties
 server.port=8080
 server.host=0.0.0.0
